chore: Refactored RowSorting into folder with base, types, and utils
